requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
pandas==2.2.3
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.1.0
requests==2.32.3
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
